Added "setSelection" method to code, tests, and documentation.
The "setSelection" method allows an explicit list of options to be set.
Deselected options in the provided list will be selected and selection
options that are not in the provided list be deselected.  Internally, it
delegates to the "select" and "deselect" methods and gives the caller
the same option for firing the onChange notification.

The testing suite has been updated to validate the new feature.  The
tests make sure that the selection changes as it is supposed to, the
onChange notification fires appropriately, and doesn't fire when
requested not to do so.

Finally, the documentation page has been updated to show how to use this
feature with working and tested examples that are based off of the other
ones already provided.