Fixes an issue when calling selectAll on a filtered result
In a dropdownlist of Contract numbers, the selectAll function on a filtered list would select all visible items plus the n th item ("n" being the filtered text). For example, if I filter on "14" and click on selectAll, I would get 114, 214, 814 and 073 (the 14th item in the list) selected.
I figured that since the deselectAll didn't have the same behavior, there was probably a small difference between them. I found it, corrected it in my project and it works properly now.